requirements.txt
pandas>=2.0.0
requests>=2.31.0
matplotlib>=3.7.0
certifi>=2023.7.22
pyyaml>=6.0
pyarrow>=12.0.0
urllib3>=2.0.0
